feat(indexing): Enhance InMemoryIndexMaintainer to support incremental updates and improved contribution management
feat(database): Update InMemoryDatabaseExecutionOrchestrator to handle previous snapshots during index updates

fix(tests): Add comprehensive tests for incremental updates and ensure query consistency across updates and rebuilds

+++ Assets/Scripts/VM/Debug/Lsp/Database/Tasks/PassThroughDatabaseTaskPlanner.cs
@@ -45,7 +45,7 @@ namespace FFVM.Debug.Lsp.Database
 					null),
 			};
 
-			int version = snapshot?.Version ?? CodeDatabaseSnapshot.Empty().Version;
+			long version = snapshot?.Version ?? CodeDatabaseSnapshot.Empty().Version;
 			return new DatabaseTaskPlan(planId, commandId, version, tasks, DateTime.UtcNow);
 		}
 	}
